LPS-178843 Migrate odata to typescript
