+++ modules/apps/segments/segments-web/src/main/resources/META-INF/resources/js/components/criteria_builder/CriteriaRowReadable.es.js
@@ -16,7 +16,7 @@ import {PropTypes} from 'prop-types';
 import React, {Component} from 'react';
 
 import {PROPERTY_TYPES} from '../../utils/constants';
-import {unescapeSingleQuotes} from '../../utils/odata.es';
+import {unescapeSingleQuotes} from '../../utils/odata';
 import {dateToInternationalHuman} from '../../utils/utils.es';
 
 class CriteriaRowReadable extends Component {
