+++ modules/apps/segments/segments-web/test/js/utils/odata.es.js
@@ -12,7 +12,7 @@
  * details.
  */
 
-import * as ODataUtil from '../../../src/main/resources/META-INF/resources/js/utils/odata.es';
+import * as ODataUtil from '../../../src/main/resources/META-INF/resources/js/utils/odata';
 import * as Utils from '../../../src/main/resources/META-INF/resources/js/utils/utils.es';
 import {mockCriteria, mockCriteriaNested} from '../mockData';
