+++ modules/apps/segments/segments-web/types/src/main/resources/META-INF/resources/js/utils/constants.d.ts
@@ -19,6 +19,7 @@ export declare const CONJUNCTIONS: {
 	readonly AND: 'and';
 	readonly OR: 'or';
 };
+export declare type Conjunction = typeof CONJUNCTIONS[keyof typeof CONJUNCTIONS];
 export declare const FUNCTIONAL_OPERATORS: {
 	readonly CONTAINS: 'contains';
 };
@@ -47,6 +48,7 @@ export declare const PROPERTY_TYPES: {
 	readonly INTEGER: 'integer';
 	readonly STRING: 'string';
 };
+export declare type PropertyType = typeof PROPERTY_TYPES[keyof typeof PROPERTY_TYPES];
 export declare const SUPPORTED_CONJUNCTIONS: readonly [
 	{
 		readonly label: string;
